pytdl/requirements.txt
Pillow==9.4.0
PyDrive==1.3.1
PyQt5==5.15.7
pytube==12.1.2
requests==2.25.1
